refactor: [US-050] - Fix dna package imports, add dna_analyser shim, extract dna/utils.py
