+++ core/tests/test_math_accuracy.py
@@ -511,7 +511,7 @@ class IsbnNormalizationTests(TestCase):
 
         from unittest.mock import patch
 
-        with patch("core.services.dna_analyser.generate_vibe_with_llm", return_value=["a", "b"]), \
+        with patch("core.services.dna.generate_vibe_with_llm", return_value=["a", "b"]), \
              patch("core.tasks.enrich_book_task.delay"):
             calculate_full_dna(goodreads_csv, user=user)
             calculate_full_dna(storygraph_csv, user=user)
